change: rename useFetchConstants directory to useConstants and update imports
