+++ src/hooks/useConstants/index.ts
@@ -1,0 +1,4 @@
+import useConstants from 'ambire-common/src/hooks/useConstants/useConstants'
+
+export * from 'ambire-common/src/hooks/useConstants/types'
+export default useConstants
