--- src/hooks/useFetchConstants/index.ts
@@ -1,4 +1,0 @@
-import useConstants from './useConstants'
-
-export * from './types'
-export default useConstants
