+++ src/services/humanReadableTransactions/transactionSummary.ts
@@ -1,8 +1,9 @@
 // TODO: add types
 // @ts-nocheck
 
+import { HumanizerInfoType } from 'ambire-common/src/hooks/useConstants'
+
 import networks from '../../constants/networks'
-import { HumanizerInfoType } from '../../hooks/useFetchConstants'
 import humanizers from '../humanizers'
 import { getName, nativeToken } from './humanReadableTransactions'
