+++ src/services/humanizers/YearnTesseractVault.ts
@@ -1,9 +1,9 @@
 // TODO: add types
 // @ts-nocheck
 
+import { HumanizerInfoType } from 'ambire-common/src/hooks/useConstants'
 import { Interface } from 'ethers/lib/utils'
 
-import { HumanizerInfoType } from '../../hooks/useFetchConstants'
 import { token } from '../humanReadableTransactions'
 
 const vaultNames = { ethereum: 'Yearn', polygon: 'Tesseract' }
