+++ src/services/humanizers/index.ts
@@ -1,5 +1,6 @@
 // Types
-import { HumanizerInfoType } from '../../hooks/useFetchConstants'
+import { HumanizerInfoType } from 'ambire-common/src/hooks/useConstants'
+
 import AaveLendingPoolV2 from './AaveLendingPoolV2'
 import AaveWethGatewayV2 from './AaveWethGatewayV2'
 import AmbireBatcher from './AmbireBatcher'
